REFACTOR: 프로젝트 구조 리팩토링 - 공통 config (#52)
- 공통 config 파일 정리
